Criação de consulta e correção de regra de negócio

+++ src/main/java/br/gov/sp/fatec/springbootprodutora/service/SegurancaService.java
@@ -6,5 +6,5 @@ import br.gov.sp.fatec.springbootprodutora.entity.Usuario;
 public interface SegurancaService {
     public Usuario criaUsuario(String nome, String senha, String autorizacao);
 
-    public Filme criaFilme(String nome, int ano, Float duracao, String descricao, String diretor, String ator, String duble);
+    public Filme criaFilme(String nome, Long ano, Float duracao, String descricao, String diretor, String ator, String duble);
 }
